Update README.md
Hotel-Booking-Analysis
AlmaBetter Verfied Project - AlmaBetter
all.jpg

📖Introduction
This data set contains booking information for a city hotel and a resort hotel, and includes information such as when the booking was made, length of stay, the number of adults, children, and/or babies, and the number of available parking spaces, among other things. All personally identifying information has been removed from the data
📖Problem Statements
Have you ever wondered when the best time of year to book a hotel room is? Or the optimal length of stay in order to get the best daily rate? What if you wanted to predict whether or not a hotel was likely to receive a disproportionately high number of special requests? This hotel booking dataset can help you explore those questions!
📖Conclusion
1)In all these processes we get to know that city hotels have more bookings than resort hotel might be because resort hotels are on costlier side or at longer distance on other has on other hand. city hotels are less costly and situated near at Railway stations and Airports. 27 % of bookings were cancelled, only 3.9% guests were repeated.
2)Bookings across year is higher for city hotel.

3)City hotel booking is increasing from January to June then decrease booking.

4)Maximum number of guests were from Portugal.

5)Room type ‘A’ is most preferred by the guests